requirements.txt
django
fastapi
psycopg[binary]
uvicorn
djangorestframework
pandas
requests
